+++ tests/flexmock_pytest_test.py
@@ -11,7 +11,10 @@ from tests.flexmock_test import assert_raises
 
 def test_module_level_test_for_pytest():
     flexmock(foo="bar").should_receive("foo").once()
-    assert_raises(MethodCallError, flexmock_teardown)
+    with assert_raises(
+        MethodCallError, "foo() expected to be called exactly 1 time, called 0 times"
+    ):
+        flexmock_teardown()
 
 
 @pytest.fixture()
@@ -26,14 +29,20 @@ def test_runtest_hook_with_fixture_for_pytest(runtest_hook_fixture):
 class TestForPytest(flexmock_test.RegularClass):
     def test_class_level_test_for_pytest(self):
         flexmock(foo="bar").should_receive("foo").once()
-        assert_raises(MethodCallError, flexmock_teardown)
+        with assert_raises(
+            MethodCallError, "foo() expected to be called exactly 1 time, called 0 times"
+        ):
+            flexmock_teardown()
 
 
 class TestUnittestClass(flexmock_test.TestFlexmockUnittest):
     def test_unittest(self):
         mocked = flexmock(a=2)
         mocked.should_receive("a").once()
-        assert_raises(MethodCallError, flexmock_teardown)
+        with assert_raises(
+            MethodCallError, "a() expected to be called exactly 1 time, called 0 times"
+        ):
+            flexmock_teardown()
 
 
 class TestFailureOnException: