+++ tests/flexmock_test.py
@@ -5,11 +5,14 @@ import random
 import re
 import sys
 import unittest
+from contextlib import contextmanager
+from typing import Type, Union
 
 from flexmock.api import (
     AT_LEAST,
     AT_MOST,
     EXACTLY,
+    RE_TYPE,
     UPDATED_ATTRS,
     CallOrderError,
     ExceptionClassError,
@@ -50,15 +53,37 @@ class SomeClass:
         pass
 
 
-def assert_raises(exception, method, *kargs, **kwargs):
+@contextmanager
+def assert_raises(expected_exception: Type[BaseException], match: Union[RE_TYPE, str, None]):
+    """Assert that code raises the correct exception with a correct error message.
+
+    Args:
+        expected_exception: Type of the expected exception.
+        match: String or pattern to match the error message against. Use None
+          to skip error message checking.
+    """
     try:
-        method(*kargs, **kwargs)
-    except exception:
-        return
-    except Exception:
-        instance = sys.exc_info()[1]
-        print("%s" % instance)
-    raise Exception("%s not raised" % exception.__name__)
+        yield
+    except Exception as raised_exception:
+        if not isinstance(raised_exception, expected_exception):
+            raise AssertionError(
+                f"Expected exception '{type(expected_exception)}' "
+                f"but '{type(raised_exception)}' was raised"
+            ) from raised_exception
+        if match is not None:
+            fail = False
+            if isinstance(match, RE_TYPE):
+                fail = not match.search(str(raised_exception))
+                match = match.pattern
+            else:
+                fail = str(raised_exception) != str(match)
+            if fail:
+                raise AssertionError(
+                    f"Expected error message:\n{'-'*39}\n'{str(match)}'\n"
+                    f"\nBut got:\n\n'{str(raised_exception)}'\n{'-'*39}\n"
+                ) from raised_exception
+    else:
+        raise AssertionError(f"Exception '{expected_exception.__name__}' not raised")
 
 
 def assert_equal(expected, received, msg=""):
@@ -103,7 +128,7 @@ class RegularClass:
             def bar(self):
                 return "bar"
 
-        flexmock(Foo).should_receive(u"bar").and_return("mocked_bar")
+        flexmock(Foo).should_receive("bar").and_return("mocked_bar")
 
         foo = Foo()
         assert_equal("mocked_bar", foo.bar())
@@ -173,7 +198,10 @@ class RegularClass:
         mock = flexmock(name="temp")
         mock.should_receive("method_foo").times(1)
         expectation = FlexmockContainer.get_flexmock_expectation(mock, "method_foo")
-        assert_raises(MethodCallError, expectation.verify)
+        with assert_raises(
+            MethodCallError, "method_foo() expected to be called exactly 1 time, called 0 times"
+        ):
+            expectation.verify()
         mock.method_foo()
         expectation.verify()
 
@@ -184,7 +212,8 @@ class RegularClass:
             pass
 
         mock.should_receive("method_foo").and_raise(FakeException)
-        assert_raises(FakeException, mock.method_foo)
+        with assert_raises(FakeException, ""):
+            mock.method_foo()
         assert_equal(
             1,
             FlexmockContainer.get_flexmock_expectation(mock, "method_foo").times_called,
@@ -199,7 +228,8 @@ class RegularClass:
                 pass
 
         mock.should_receive("method_foo").and_raise(FakeException(1, arg2=2))
-        assert_raises(FakeException, mock.method_foo)
+        with assert_raises(FakeException, "1"):
+            mock.method_foo()
         assert_equal(
             1,
             FlexmockContainer.get_flexmock_expectation(mock, "method_foo").times_called,
@@ -214,7 +244,8 @@ class RegularClass:
                 pass
 
         mock.should_receive("method_foo").and_raise(FakeException, 1, arg2=2)
-        assert_raises(FakeException, mock.method_foo)
+        with assert_raises(FakeException, "1"):
+            mock.method_foo()
         assert_equal(
             1,
             FlexmockContainer.get_flexmock_expectation(mock, "method_foo").times_called,
@@ -232,7 +263,15 @@ class RegularClass:
     def test_flexmock_should_fail_to_match_exactly_no_args_when_calling_with_args(self):
         mock = flexmock()
         mock.should_receive("method_foo").with_args()
-        assert_raises(MethodSignatureError, mock.method_foo, "baz")
+        with assert_raises(
+            MethodSignatureError,
+            (
+                "Arguments for call method_foo did not match expectations:\n"
+                '  Received call:\tmethod_foo("baz")\n'
+                "  Expected call[1]:\tmethod_foo()"
+            ),
+        ):
+            mock.method_foo("baz")
 
     def test_flexmock_should_match_exactly_no_args(self):
         class Foo:
@@ -313,7 +352,16 @@ class RegularClass:
         mock.should_receive("method_foo").with_args(int).and_return("got an int")
         assert_equal("got a string", mock.method_foo("string!"))
         assert_equal("got an int", mock.method_foo(23))
-        assert_raises(MethodSignatureError, mock.method_foo, 2.0)
+        with assert_raises(
+            MethodSignatureError,
+            (
+                "Arguments for call method_foo did not match expectations:\n"
+                "  Received call:\tmethod_foo(2.0)\n"
+                "  Expected call[1]:\tmethod_foo(<class 'int'>)\n"
+                "  Expected call[2]:\tmethod_foo(<class 'str'>)"
+            ),
+        ):
+            mock.method_foo(2.0)
 
     def test_with_args_should_work_with_builtin_c_methods(self):
         flexmock(sys.stdout).should_call("write")  # set fall-through
@@ -340,7 +388,16 @@ class RegularClass:
 
         mock.should_receive("method_foo").with_args(Foo).and_return("got a Foo")
         assert_equal("got a Foo", mock.method_foo(Foo()))
-        assert_raises(MethodSignatureError, mock.method_foo, 1)
+        with assert_raises(
+            MethodSignatureError,
+            (
+                "Arguments for call method_foo did not match expectations:\n"
+                "  Received call:\tmethod_foo(1)\n"
+                "  Expected call[1]:\tmethod_foo(<class 'tests.flexmock_test.RegularClass.test"
+                "_flexmock_should_match_expectations_against_user_defined_classes.<locals>.Foo'>)"
+            ),
+        ):
+            mock.method_foo(1)
 
     def test_flexmock_configures_global_mocks_dict(self):
         mock = flexmock(name="temp")
@@ -352,7 +409,11 @@ class RegularClass:
     def test_flexmock_teardown_verifies_mocks(self):
         mock = flexmock(name="temp")
         mock.should_receive("verify_expectations").times(1)
-        assert_raises(MethodCallError, self._tear_down)
+        with assert_raises(
+            MethodCallError,
+            "verify_expectations() expected to be called exactly 1 time, called 0 times",
+        ):
+            self._tear_down()
 
     def test_flexmock_teardown_does_not_verify_stubs(self):
         mock = flexmock(name="temp")
@@ -459,7 +520,10 @@ class RegularClass:
         expectation = FlexmockContainer.get_flexmock_expectation(mock, "method_foo")
         assert_equal(AT_LEAST, expectation._modifier)
         mock.method_foo()
-        assert_raises(MethodCallError, self._tear_down)
+        with assert_raises(
+            MethodCallError, "method_foo() expected to be called at least 2 times, called 1 time"
+        ):
+            self._tear_down()
 
     def test_flexmock_respects_at_least_when_called_requested_number(self):
         mock = flexmock(name="temp")
@@ -500,21 +564,30 @@ class RegularClass:
         expectation = FlexmockContainer.get_flexmock_expectation(mock, "method_foo")
         assert_equal(AT_MOST, expectation._modifier)
         mock.method_foo()
-        assert_raises(MethodCallError, mock.method_foo)
+        with assert_raises(
+            MethodCallError, "method_foo() expected to be called at most 1 time, called 2 times"
+        ):
+            mock.method_foo()
 
     def test_flexmock_treats_once_as_times_one(self):
         mock = flexmock(name="temp")
         mock.should_receive("method_foo").and_return("value_bar").once()
         expectation = FlexmockContainer.get_flexmock_expectation(mock, "method_foo")
         assert_equal(1, expectation._expected_calls[EXACTLY])
-        assert_raises(MethodCallError, self._tear_down)
+        with assert_raises(
+            MethodCallError, "method_foo() expected to be called exactly 1 time, called 0 times"
+        ):
+            self._tear_down()
 
     def test_flexmock_treats_twice_as_times_two(self):
         mock = flexmock(name="temp")
         mock.should_receive("method_foo").twice().and_return("value_bar")
         expectation = FlexmockContainer.get_flexmock_expectation(mock, "method_foo")
         assert_equal(2, expectation._expected_calls[EXACTLY])
-        assert_raises(MethodCallError, self._tear_down)
+        with assert_raises(
+            MethodCallError, "method_foo() expected to be called exactly 2 times, called 0 times"
+        ):
+            self._tear_down()
 
     def test_flexmock_works_with_never_when_true(self):
         mock = flexmock(name="temp")
@@ -526,7 +599,10 @@ class RegularClass:
     def test_flexmock_works_with_never_when_false(self):
         mock = flexmock(name="temp")
         mock.should_receive("method_foo").and_return("value_bar").never()
-        assert_raises(MethodCallError, mock.method_foo)
+        with assert_raises(
+            MethodCallError, "method_foo() expected to be called exactly 0 times, called 1 time"
+        ):
+            mock.method_foo()
 
     def test_flexmock_get_flexmock_expectation_should_work_with_args(self):
         mock = flexmock(name="temp")
@@ -565,7 +641,12 @@ class RegularClass:
                 return a
 
         instance = WithCall()
-        assert_raises(FlexmockError, flexmock(instance).should_call, "__call__")
+
+        with assert_raises(
+            FlexmockError,
+            re.compile(r".+<locals>\.WithCall object at 0x.+> does not have attribute '__call__'"),
+        ):
+            flexmock(instance).should_call("__call__")
 
     def test_should_call_on_class_mock(self):
         class User:
@@ -582,14 +663,20 @@ class RegularClass:
         user1 = User()
         user2 = User()
         flexmock(User).should_call("foo").once()
-        assert_raises(MethodCallError, self._tear_down)
+        with assert_raises(
+            MethodCallError, "foo() expected to be called exactly 1 time, called 0 times"
+        ):
+            self._tear_down()
         flexmock(User).should_call("foo").twice()
         assert_equal("class", user1.foo())
         assert_equal("class", user2.foo())
 
         # Access instance attributes
         flexmock(User).should_call("bar").once()
-        assert_raises(MethodCallError, self._tear_down)
+        with assert_raises(
+            MethodCallError, "bar() expected to be called exactly 1 time, called 0 times"
+        ):
+            self._tear_down()
         flexmock(User).should_call("bar").twice()
         assert_equal("value", user1.bar())
         assert_equal("value", user2.bar())
@@ -689,7 +776,10 @@ class RegularClass:
         flexmock(User)
         Group.should_receive("method1").once()
         User.should_receive("method2").once()
-        assert_raises(MethodCallError, self._tear_down)
+        with assert_raises(
+            MethodCallError, "method1() expected to be called exactly 1 time, called 0 times"
+        ):
+            self._tear_down()
         for method in UPDATED_ATTRS:
             assert method not in dir(Group)
         for method in UPDATED_ATTRS:
@@ -721,12 +811,18 @@ class RegularClass:
 
         group = Group()
         flexmock(group).should_call("method1").at_least().once()
-        assert_raises(MethodCallError, self._tear_down)
+        with assert_raises(
+            MethodCallError, "method1() expected to be called at least 1 time, called 0 times"
+        ):
+            self._tear_down()
         flexmock(group)
         group.should_call("method2").with_args("a").once()
         group.should_receive("method2").with_args("not a")
         group.method2("not a")
-        assert_raises(MethodCallError, self._tear_down)
+        with assert_raises(
+            MethodCallError, 'method2(a="a") expected to be called exactly 1 time, called 0 times'
+        ):
+            self._tear_down()
 
     def test_flexmock_doesnt_error_on_properly_ordered_expectations(self):
         class Foo:
@@ -763,7 +859,8 @@ class RegularClass:
         flexmock(foo)
         foo.should_receive("method1").with_args("a").ordered()
         foo.should_receive("method1").with_args("b").ordered()
-        assert_raises(CallOrderError, foo.method1, "b")
+        with assert_raises(CallOrderError, 'method1("b") called before method1(a="a")'):
+            foo.method1("b")
 
     def test_flexmock_should_accept_multiple_return_values(self):
         class Foo:
@@ -821,9 +918,11 @@ class RegularClass:
         foo = Foo()
         flexmock(foo).should_receive("method1").and_return(1).and_raise(Exception)
         assert_equal(1, foo.method1())
-        assert_raises(Exception, foo.method1)
+        with assert_raises(Exception, ""):
+            foo.method1()
         assert_equal(1, foo.method1())
-        assert_raises(Exception, foo.method1)
+        with assert_raises(Exception, ""):
+            foo.method1()
 
     def test_flexmock_should_match_types_on_multiple_arguments(self):
         class Foo:
@@ -833,11 +932,38 @@ class RegularClass:
         foo = Foo()
         flexmock(foo).should_receive("method1").with_args(str, int).and_return("ok")
         assert_equal("ok", foo.method1("some string", 12))
-        assert_raises(MethodSignatureError, foo.method1, 12, 32)
+        with assert_raises(
+            MethodSignatureError,
+            (
+                "Arguments for call method1 did not match expectations:\n"
+                "  Received call:\tmethod1(12, 32)\n"
+                "  Expected call[1]:\tmethod1(a=<class 'str'>, b=<class 'int'>)"
+            ),
+        ):
+            foo.method1(12, 32)
         flexmock(foo).should_receive("method1").with_args(str, int).and_return("ok")
-        assert_raises(MethodSignatureError, foo.method1, 12, "some string")
+        with assert_raises(
+            MethodSignatureError,
+            (
+                "Arguments for call method1 did not match expectations:\n"
+                '  Received call:\tmethod1(12, "some string")\n'
+                "  Expected call[1]:\tmethod1(a=<class 'str'>, b=<class 'int'>)\n"
+                "  Expected call[2]:\tmethod1(a=<class 'str'>, b=<class 'int'>)"
+            ),
+        ):
+            foo.method1(12, "some string")
         flexmock(foo).should_receive("method1").with_args(str, int).and_return("ok")
-        assert_raises(MethodSignatureError, foo.method1, "string", 12, 14)
+        with assert_raises(
+            MethodSignatureError,
+            (
+                "Arguments for call method1 did not match expectations:\n"
+                '  Received call:\tmethod1("string", 12, 14)\n'
+                "  Expected call[1]:\tmethod1(a=<class 'str'>, b=<class 'int'>)\n"
+                "  Expected call[2]:\tmethod1(a=<class 'str'>, b=<class 'int'>)\n"
+                "  Expected call[3]:\tmethod1(a=<class 'str'>, b=<class 'int'>)"
+            ),
+        ):
+            foo.method1("string", 12, 14)
 
     def test_flexmock_should_match_types_on_multiple_arguments_generic(self):
         class Foo:
@@ -850,9 +976,28 @@ class RegularClass:
         assert_equal("ok", foo.method1((1,), None, 12))
         assert_equal("ok", foo.method1(12, 14, []))
         assert_equal("ok", foo.method1("some string", "another one", False))
-        assert_raises(MethodSignatureError, foo.method1, "string", 12)
+        with assert_raises(
+            MethodSignatureError,
+            (
+                "Arguments for call method1 did not match expectations:\n"
+                '  Received call:\tmethod1("string", 12)\n'
+                "  Expected call[1]:\tmethod1(a=<class 'object'>, "
+                "b=<class 'object'>, c=<class 'object'>)"
+            ),
+        ):
+            foo.method1("string", 12)  # pylint: disable=no-value-for-parameter
+        self._tear_down()
         flexmock(foo).should_receive("method1").with_args(object, object, object).and_return("ok")
-        assert_raises(MethodSignatureError, foo.method1, "string", 12, 13, 14)
+        with assert_raises(
+            MethodSignatureError,
+            (
+                "Arguments for call method1 did not match expectations:\n"
+                '  Received call:\tmethod1("string", 12, 13, 14)\n'
+                "  Expected call[1]:\tmethod1(a=<class 'object'>, "
+                "b=<class 'object'>, c=<class 'object'>)"
+            ),
+        ):
+            foo.method1("string", 12, 13, 14)
 
     def test_flexmock_should_match_types_on_multiple_arguments_classes(self):
         class Foo:
@@ -866,9 +1011,26 @@ class RegularClass:
         bar = Bar()
         flexmock(foo).should_receive("method1").with_args(object, Bar).and_return("ok")
         assert_equal("ok", foo.method1("some string", bar))
-        assert_raises(MethodSignatureError, foo.method1, bar, "some string")
+        with assert_raises(
+            MethodSignatureError,
+            re.compile(
+                "Arguments for call method1 did not match expectations:\n"
+                r'  Received call:\tmethod1\(.+\.<locals>\.Bar object at 0x.+>, "some string"\)\n'
+                r"  Expected call\[1\]:\tmethod1\(a=<class 'object'>, b=<class.+\.<locals>\.Bar'>\)"
+            ),
+        ):
+            foo.method1(bar, "some string")
+        self._tear_down()
         flexmock(foo).should_receive("method1").with_args(object, Bar).and_return("ok")
-        assert_raises(MethodSignatureError, foo.method1, 12, "some string")
+        with assert_raises(
+            MethodSignatureError,
+            re.compile(
+                "Arguments for call method1 did not match expectations:\n"
+                r'  Received call:\tmethod1\(12, "some string"\)\n'
+                r"  Expected call\[1\]:\tmethod1\(a=<class 'object'>, b=<class.+\.<locals>\.Bar'>\)"
+            ),
+        ):
+            foo.method1(12, "some string")
 
     def test_flexmock_should_match_keyword_arguments(self):
         class Foo:
@@ -881,11 +1043,37 @@ class RegularClass:
         foo.method1(1, arg3=3, arg2=2)
         self._tear_down()
         flexmock(foo).should_receive("method1").with_args(1, arg3=3, arg2=2)
-        assert_raises(MethodSignatureError, foo.method1, arg2=2, arg3=3)
+        with assert_raises(
+            MethodSignatureError,
+            (
+                "Arguments for call method1 did not match expectations:\n"
+                "  Received call:\tmethod1(arg2=2, arg3=3)\n"
+                "  Expected call[1]:\tmethod1(arg3=3, arg2=2, a=1)"
+            ),
+        ):
+            foo.method1(arg2=2, arg3=3)  # pylint: disable=no-value-for-parameter
+        self._tear_down()
         flexmock(foo).should_receive("method1").with_args(1, arg3=3, arg2=2)
-        assert_raises(MethodSignatureError, foo.method1, 1, arg2=2, arg3=4)
+        with assert_raises(
+            MethodSignatureError,
+            (
+                "Arguments for call method1 did not match expectations:\n"
+                "  Received call:\tmethod1(1, arg2=2, arg3=4)\n"
+                "  Expected call[1]:\tmethod1(arg3=3, arg2=2, a=1)"
+            ),
+        ):
+            foo.method1(1, arg2=2, arg3=4)
+        self._tear_down()
         flexmock(foo).should_receive("method1").with_args(1, arg3=3, arg2=2)
-        assert_raises(MethodSignatureError, foo.method1, 1)
+        with assert_raises(
+            MethodSignatureError,
+            (
+                "Arguments for call method1 did not match expectations:\n"
+                "  Received call:\tmethod1(1)\n"
+                "  Expected call[1]:\tmethod1(arg3=3, arg2=2, a=1)"
+            ),
+        ):
+            foo.method1(1)
 
     def test_flexmock_should_call_should_match_keyword_arguments(self):
         class Foo:
@@ -1057,7 +1245,15 @@ class RegularClass:
 
         user = User()
         flexmock(user).should_call("get_stuff").and_raise(FakeException, "2", "1")
-        assert_raises(ExceptionMessageError, user.get_stuff)
+        with assert_raises(
+            ExceptionMessageError,
+            (
+                "Error message mismatch with raised FakeException:\n"
+                "  Expected message:\n\t'1, 2'\n"
+                "  Received message:\n\t'2, 1'"
+            ),
+        ):
+            user.get_stuff()
 
     def test_flexmock_should_verify_spy_matches_exception_regexp(self):
         class User:
@@ -1076,7 +1272,15 @@ class RegularClass:
 
         user = User()
         flexmock(user).should_call("get_stuff").and_raise(Exception, re.compile("^asdf"))
-        assert_raises(ExceptionMessageError, user.get_stuff)
+        with assert_raises(
+            ExceptionMessageError,
+            (
+                "Error message mismatch with raised Exception:\n"
+                "  Expected pattern:\n\t/^asdf/\n"
+                "  Received message:\n\t'123asdf345'"
+            ),
+        ):
+            user.get_stuff()
 
     def test_flexmock_should_blow_up_on_wrong_spy_exception_type(self):
         class User:
@@ -1085,7 +1289,15 @@ class RegularClass:
 
         user = User()
         flexmock(user).should_call("get_stuff").and_raise(MethodCallError)
-        assert_raises(ExceptionClassError, user.get_stuff)
+        with assert_raises(
+            ExceptionClassError,
+            (
+                "Raised exception for call get_stuff did not match expectation:\n"
+                "  Expected:\t<class 'flexmock.exceptions.MethodCallError'>\n"
+                "  Raised:\t<class 'flexmock.exceptions.CallOrderError'>"
+            ),
+        ):
+            user.get_stuff()
 
     def test_flexmock_should_match_spy_exception_parent_type(self):
         class User:
@@ -1106,9 +1318,25 @@ class RegularClass:
 
         user = User()
         flexmock(user).should_call("get_stuff").and_return("other", "stuff")
-        assert_raises(MethodSignatureError, user.get_stuff)
+        with assert_raises(
+            MethodSignatureError,
+            (
+                "Returned values for call get_stuff did not match expectation:\n"
+                "  Expected:\t('other', 'stuff')\n"
+                "  Returned:\t('real', 'stuff')"
+            ),
+        ):
+            user.get_stuff()
         flexmock(user).should_call("get_more_stuff").and_return()
-        assert_raises(MethodSignatureError, user.get_more_stuff)
+        with assert_raises(
+            MethodSignatureError,
+            (
+                "Returned values for call get_more_stuff did not match expectation:\n"
+                "  Expected:\tNone\n"
+                "  Returned:\t('other', 'stuff')"
+            ),
+        ):
+            user.get_more_stuff()
 
     def test_flexmock_should_mock_same_class_twice(self):
         class Foo:
@@ -1168,7 +1396,15 @@ class RegularClass:
         else:
             mod = sys.modules["__main__"]
         flexmock(mod).should_receive("module_level_function").with_args(1, args="expected")
-        assert_raises(FlexmockError, module_level_function, 1, args="not expected")
+        with assert_raises(
+            FlexmockError,
+            (
+                "Arguments for call module_level_function did not match expectations:\n"
+                '  Received call:\tmodule_level_function(1, args="not expected")\n'
+                '  Expected call[1]:\tmodule_level_function(args="expected", some=1)'
+            ),
+        ):
+            module_level_function(1, args="not expected")
 
     def test_flexmock_should_support_mocking_classes_as_functions(self):
         if "tests.flexmock_test" in sys.modules:
@@ -1229,17 +1465,50 @@ class RegularClass:
         flexmock(foo).should_call("bar").and_return("bar").once()
         flexmock(foo).should_call("baz").and_return("bar").once()
         flexmock(foo).should_call("quux").and_return("bar").once()
-        assert_raises(FlexmockError, foo.foo)
-        assert_raises(FlexmockError, foo.bar)
-        assert_raises(FlexmockError, foo.baz)
-        assert_raises(FlexmockError, foo.quux)
+        with assert_raises(
+            FlexmockError,
+            (
+                "Returned values for call foo did not match expectation:\n"
+                "  Expected:\t'bar'\n"
+                "  Returned:\tNone"
+            ),
+        ):
+            foo.foo()
+        with assert_raises(
+            FlexmockError,
+            (
+                "Returned values for call bar did not match expectation:\n"
+                "  Expected:\t'bar'\n"
+                "  Returned:\tFalse"
+            ),
+        ):
+            foo.bar()
+        with assert_raises(
+            FlexmockError,
+            (
+                "Returned values for call baz did not match expectation:\n"
+                "  Expected:\t'bar'\n"
+                "  Returned:\t[]"
+            ),
+        ):
+            foo.baz()
+        with assert_raises(
+            FlexmockError,
+            (
+                "Returned values for call quux did not match expectation:\n"
+                "  Expected:\t'bar'\n"
+                "  Returned:\t''"
+            ),
+        ):
+            foo.quux()
 
     def test_new_instances_should_blow_up_on_should_receive(self):
         class User:
             pass
 
         mock = flexmock(User).new_instances(None).mock
-        assert_raises(FlexmockError, mock.should_receive, "foo")
+        with assert_raises(FlexmockError, "User does not have attribute 'foo'"):
+            mock.should_receive("foo")
 
     def test_should_call_alias_should_create_a_spy(self):
         class Foo:
@@ -1248,14 +1517,23 @@ class RegularClass:
 
         foo = Foo()
         flexmock(foo).should_call("get_stuff").and_return("yay").once()
-        assert_raises(MethodCallError, self._tear_down)
+        with assert_raises(
+            MethodCallError, "get_stuff() expected to be called exactly 1 time, called 0 times"
+        ):
+            self._tear_down()
 
     def test_flexmock_should_fail_mocking_nonexistent_methods(self):
         class User:
             pass
 
         user = User()
-        assert_raises(FlexmockError, flexmock(user).should_receive, "nonexistent")
+        with assert_raises(
+            FlexmockError,
+            re.compile(
+                r"<.+\.<locals>\.User object at 0x.+> does not have attribute 'nonexistent'"
+            ),
+        ):
+            flexmock(user).should_receive("nonexistent")
 
     def test_flexmock_should_not_explode_on_unicode_formatting(self):
         formatted = _format_args("method", {"kargs": (chr(0x86C7),), "kwargs": {}})
@@ -1302,13 +1580,31 @@ class RegularClass:
         flexmock(foo).should_call("method").with_args("foo").once()
         flexmock(foo).should_call("method").with_args("bar").once()
         foo.method("foo")
-        assert_raises(MethodCallError, self._tear_down)
+        with assert_raises(
+            MethodCallError,
+            'method(arg="bar") expected to be called exactly 1 time, called 0 times',
+        ):
+            self._tear_down()
 
     def test_should_give_reasonable_error_for_builtins(self):
-        assert_raises(MockBuiltinError, flexmock, object)
+        with assert_raises(
+            MockBuiltinError,
+            (
+                "Python does not allow you to mock builtin objects or modules. "
+                "Consider wrapping it in a class you can mock instead"
+            ),
+        ):
+            flexmock(object)
 
     def test_should_give_reasonable_error_for_instances_of_builtins(self):
-        assert_raises(MockBuiltinError, flexmock, object())
+        with assert_raises(
+            MockBuiltinError,
+            (
+                "Python does not allow you to mock instances of builtin objects. "
+                "Consider wrapping it in a class you can mock instead"
+            ),
+        ):
+            flexmock(object())
 
     def test_mock_chained_method_calls_works_with_one_level(self):
         class Foo:
@@ -1339,7 +1635,10 @@ class RegularClass:
         assert_equal("bar", foo.method1().method2("foo"))
         self._tear_down()
         flexmock(foo).should_receive("method1.method2").with_args("foo").and_return("bar").once()
-        assert_raises(MethodCallError, self._tear_down)
+        with assert_raises(
+            MethodCallError, 'method2("foo") expected to be called exactly 1 time, called 0 times'
+        ):
+            self._tear_down()
 
     def test_mock_chained_method_calls_works_with_more_than_one_level(self):
         class Baz:
@@ -1376,7 +1675,8 @@ class RegularClass:
 
         foo = Foo()
         expectation = flexmock(foo).should_receive("method").replace_with(lambda x: x == 5)
-        assert_raises(FlexmockError, expectation.replace_with, lambda x: x == 3)
+        with assert_raises(FlexmockError, "replace_with cannot be specified twice"):
+            expectation.replace_with(lambda x: x == 3)
 
     def test_flexmock_should_mock_the_same_method_multiple_times(self):
         class Foo:
@@ -1408,7 +1708,8 @@ class RegularClass:
 
         foo = Foo()
         flexmock(foo)
-        assert_raises(FlexmockError, foo.new_instances, "bar")
+        with assert_raises(FlexmockError, "new_instances can only be called on a class mock"):
+            foo.new_instances("bar")
 
     def test_new_instances_works_with_multiple_return_values(self):
         class Foo:
@@ -1425,7 +1726,9 @@ class RegularClass:
 
         foo = Foo()
         flexmock(foo)
-        assert_raises(FlexmockError, foo.should_receive, "should_receive")
+        for method_name in ["should_receive", "should_call", "new_instances"]:
+            with assert_raises(FlexmockError, "unable to replace flexmock methods"):
+                foo.should_receive(method_name)
 
     def test_flexmock_should_not_add_methods_if_they_already_exist(self):
         class Foo:
@@ -1501,7 +1804,15 @@ class RegularClass:
         flexmock(foo).should_receive("foo").with_args(
             re.compile("^arg1.*asdf$"), arg2=re.compile("a")
         ).and_return("mocked")
-        assert_raises(MethodSignatureError, foo.foo, "arg1somejunkasdfa", arg2="a")
+        with assert_raises(
+            MethodSignatureError,
+            (
+                "Arguments for call foo did not match expectations:\n"
+                '  Received call:\tfoo("arg1somejunkasdfa", arg2="a")\n'
+                "  Expected call[1]:\tfoo(arg2=/a/, arg1=/^arg1.*asdf$/)"
+            ),
+        ):
+            foo.foo("arg1somejunkasdfa", arg2="a")
 
     def test_arg_matching_with_regexp_fails_when_regexp_doesnt_match_kwarg(self):
         class Foo:
@@ -1512,7 +1823,15 @@ class RegularClass:
         flexmock(foo).should_receive("foo").with_args(
             re.compile("^arg1.*asdf$"), arg2=re.compile("a")
         ).and_return("mocked")
-        assert_raises(MethodSignatureError, foo.foo, "arg1somejunkasdf", arg2="b")
+        with assert_raises(
+            MethodSignatureError,
+            (
+                "Arguments for call foo did not match expectations:\n"
+                '  Received call:\tfoo("arg1somejunkasdf", arg2="b")\n'
+                "  Expected call[1]:\tfoo(arg2=/a/, arg1=/^arg1.*asdf$/)"
+            ),
+        ):
+            foo.foo("arg1somejunkasdf", arg2="b")
 
     def test_flexmock_class_returns_same_object_on_repeated_calls(self):
         class Foo:
@@ -1536,7 +1855,8 @@ class RegularClass:
         foo.should_receive("bar")
         foo.should_receive("bar").with_args("a").ordered()
         foo.should_receive("bar").with_args("b").ordered()
-        assert_raises(CallOrderError, foo.bar, "b")
+        with assert_raises(CallOrderError, 'bar("b") called before bar("a")'):
+            foo.bar("b")
 
     def test_flexmock_ordered_works_with_same_args(self):
         foo = flexmock()
@@ -1586,14 +1906,21 @@ class RegularClass:
         radio.should_receive("select_channel").once().when(lambda: radio.is_on)
         radio.should_call("adjust_volume").once().with_args(5).when(radio_is_on)
 
-        assert_raises(StateError, radio.select_channel)
-        assert_raises(StateError, radio.adjust_volume, 5)
+        with assert_raises(
+            StateError, "select_channel expected to be called when lambda: radio.is_on is True"
+        ):
+            radio.select_channel()
+        with assert_raises(
+            StateError, "adjust_volume expected to be called when radio_is_on is True"
+        ):
+            radio.adjust_volume(5)
         radio.is_on = True
         radio.select_channel()
         radio.adjust_volume(5)
 
     def test_when_parameter_should_be_callable(self):
-        assert_raises(FlexmockError, flexmock().should_receive("something").when, 1)
+        with assert_raises(FlexmockError, "when() parameter must be callable"):
+            flexmock().should_receive("something").when(1)
 
     def test_support_at_least_and_at_most_together(self):
         class Foo:
@@ -1602,12 +1929,19 @@ class RegularClass:
 
         foo = Foo()
         flexmock(foo).should_call("bar").at_least().once().at_most().twice()
-        assert_raises(MethodCallError, self._tear_down)
+        with assert_raises(
+            MethodCallError,
+            "bar() expected to be called at least 1 time and at most 2 times, called 0 times",
+        ):
+            self._tear_down()
 
         flexmock(foo).should_call("bar").at_least().once().at_most().twice()
         foo.bar()
         foo.bar()
-        assert_raises(MethodCallError, foo.bar)
+        with assert_raises(
+            MethodCallError, "bar() expected to be called at most 2 times, called 3 times"
+        ):
+            foo.bar()
 
         flexmock(foo).should_call("bar").at_least().once().at_most().twice()
         foo.bar()
@@ -1624,11 +1958,8 @@ class RegularClass:
                 pass
 
         expectation = flexmock(Foo).should_receive("bar")
-        try:
+        with assert_raises(FlexmockError, "cannot use at_least modifier twice"):
             expectation.at_least().at_least()
-            raise Exception("should not be able to specify at_least twice")
-        except FlexmockError:
-            pass
 
     def test_at_most_cannot_be_used_twice(self):
         class Foo:
@@ -1636,11 +1967,8 @@ class RegularClass:
                 pass
 
         expectation = flexmock(Foo).should_receive("bar")
-        try:
+        with assert_raises(FlexmockError, "cannot use at_most modifier twice"):
             expectation.at_most().at_most()
-            raise Exception("should not be able to specify at_most twice")
-        except FlexmockError:
-            pass
 
     def test_at_least_cannot_be_specified_until_at_most_is_set(self):
         class Foo:
@@ -1648,11 +1976,8 @@ class RegularClass:
                 pass
 
         expectation = flexmock(Foo).should_receive("bar")
-        try:
+        with assert_raises(FlexmockError, "cannot use at_most with at_least unset"):
             expectation.at_least().at_most()
-            raise Exception("should not be able to specify at_most if at_least unset")
-        except FlexmockError:
-            pass
 
     def test_at_most_cannot_be_specified_until_at_least_is_set(self):
         class Foo:
@@ -1660,11 +1985,8 @@ class RegularClass:
                 pass
 
         expectation = flexmock(Foo).should_receive("bar")
-        try:
+        with assert_raises(FlexmockError, "cannot use at_least with at_most unset"):
             expectation.at_most().at_least()
-            raise Exception("should not be able to specify at_least if at_most unset")
-        except FlexmockError:
-            pass
 
     def test_proper_reset_of_subclass_methods(self):
         class A:
@@ -1701,7 +2023,15 @@ class RegularClass:
         flexmock(s)
         s.should_call("startswith").with_args("asdf", 0, 4).ordered()
         s.should_call("endswith").ordered()
-        assert_raises(CallOrderError, s.endswith, "c")
+        with assert_raises(
+            CallOrderError,
+            re.compile(
+                r'endswith\("c"\) called before startswith'
+                # Argument names are displayed in PyPy
+                r'\((prefix=)?"asdf", (start=)?0, (end=)?4\)'
+            ),
+        ):
+            s.endswith("c")
 
     def test_fake_object_takes_properties(self):
         foo = flexmock(bar=property(lambda self: "baz"))
@@ -1767,16 +2097,26 @@ class RegularClass:
 
         foo = Foo()
         e = flexmock(foo).should_receive("bar").and_return(2)
-        assert_raises(FlexmockError, e.times, 1)
-        assert_raises(FlexmockError, e.with_args, ())
-        assert_raises(FlexmockError, e.replace_with, lambda x: x)
-        assert_raises(FlexmockError, e.and_raise, Exception)
-        assert_raises(FlexmockError, e.when, lambda x: x)
-        assert_raises(FlexmockError, e.and_yield, 1)
-        assert_raises(FlexmockError, object.__getattribute__(e, "ordered"))
-        assert_raises(FlexmockError, object.__getattribute__(e, "at_least"))
-        assert_raises(FlexmockError, object.__getattribute__(e, "at_most"))
-        assert_raises(FlexmockError, object.__getattribute__(e, "one_by_one"))
+        with assert_raises(FlexmockError, "can't use times() with attribute stubs"):
+            e.times(1)
+        with assert_raises(FlexmockError, "can't use with_args() with attribute stubs"):
+            e.with_args(())
+        with assert_raises(FlexmockError, "can't use replace_with() with attribute/property stubs"):
+            e.replace_with(lambda x: x)
+        with assert_raises(FlexmockError, "can't use and_raise() with attribute stubs"):
+            e.and_raise(Exception)
+        with assert_raises(FlexmockError, "can't use when() with attribute stubs"):
+            e.when(lambda x: x)
+        with assert_raises(FlexmockError, "can't use and_yield() with attribute stubs"):
+            e.and_yield(1)
+        with assert_raises(FlexmockError, "can't use ordered() with attribute stubs"):
+            object.__getattribute__(e, "ordered")()
+        with assert_raises(FlexmockError, "can't use at_least() with attribute stubs"):
+            object.__getattribute__(e, "at_least")()
+        with assert_raises(FlexmockError, "can't use at_most() with attribute stubs"):
+            object.__getattribute__(e, "at_most")()
+        with assert_raises(FlexmockError, "can't use one_by_one() with attribute stubs"):
+            object.__getattribute__(e, "one_by_one")()
 
     def test_and_return_defaults_to_none_with_no_arguments(self):
         foo = flexmock()
@@ -1847,11 +2187,23 @@ class RegularClass:
 
     def test_with_args_blows_up_on_too_few_args(self):
         class Foo:
+            def foo(self, a):
+                pass
+
             def bar(self, a, b, c=1):
                 pass
 
-        e = flexmock(Foo).should_receive("bar")
-        assert_raises(MethodSignatureError, e.with_args, 1)
+        mock = flexmock(Foo)
+        e1 = mock.should_receive("foo")
+        e2 = mock.should_receive("bar")
+        with assert_raises(
+            MethodSignatureError, "foo requires at least 1 argument, expectation provided 0"
+        ):
+            e1.with_args()
+        with assert_raises(
+            MethodSignatureError, "bar requires at least 2 arguments, expectation provided 1"
+        ):
+            e2.with_args(1)
 
     def test_with_args_blows_up_on_too_few_args_with_kwargs(self):
         class Foo:
@@ -1859,7 +2211,10 @@ class RegularClass:
                 pass
 
         e = flexmock(Foo).should_receive("bar")
-        assert_raises(MethodSignatureError, e.with_args, 1, c=2)
+        with assert_raises(
+            MethodSignatureError, "bar requires at least 3 arguments, expectation provided 2"
+        ):
+            e.with_args(1, c=2)
 
     def test_with_args_blows_up_on_too_many_args(self):
         class Foo:
@@ -1867,7 +2222,10 @@ class RegularClass:
                 pass
 
         e = flexmock(Foo).should_receive("bar")
-        assert_raises(MethodSignatureError, e.with_args, 1, 2, 3, 4)
+        with assert_raises(
+            MethodSignatureError, "bar requires at most 3 arguments, expectation provided 4"
+        ):
+            e.with_args(1, 2, 3, 4)
 
     def test_with_args_blows_up_on_kwarg_overlapping_positional(self):
         class Foo:
@@ -1875,7 +2233,14 @@ class RegularClass:
                 pass
 
         e = flexmock(Foo).should_receive("bar")
-        assert_raises(MethodSignatureError, e.with_args, 1, 2, 3, c=2)
+        with assert_raises(
+            MethodSignatureError, "['c'] already given as positional argument to bar"
+        ):
+            e.with_args(1, 2, 3, c=2)
+        with assert_raises(
+            MethodSignatureError, "['c', 'b'] already given as positional arguments to bar"
+        ):
+            e.with_args(1, 2, 3, c=2, b=3)
 
     def test_with_args_blows_up_on_invalid_kwarg(self):
         class Foo:
@@ -1883,7 +2248,8 @@ class RegularClass:
                 pass
 
         e = flexmock(Foo).should_receive("bar")
-        assert_raises(MethodSignatureError, e.with_args, 1, 2, d=2)
+        with assert_raises(MethodSignatureError, "d is not a valid keyword argument to bar"):
+            e.with_args(1, 2, d=2)
 
     def test_with_args_ignores_invalid_args_on_flexmock_instances(self):
         foo = flexmock(bar=lambda x: x)
@@ -2071,21 +2437,26 @@ class TestUnittestModern(TestFlexmockUnittestModern):
 
 class TestPy3Features(unittest.TestCase):
     def test_mock_kwargs_only_func_mock_all(self):
-        flexmock(some_module).should_receive("kwargs_only_func").with_args(
+        flexmock(some_module).should_receive("kwargs_only_func1").with_args(
             1, bar=2, baz=3
         ).and_return(123)
-        self.assertEqual(some_module.kwargs_only_func(1, bar=2, baz=3), 123)
+        self.assertEqual(some_module.kwargs_only_func1(1, bar=2, baz=3), 123)
 
     def test_mock_kwargs_only_func_mock_required(self):
-        flexmock(some_module).should_receive("kwargs_only_func").with_args(1, bar=2).and_return(123)
-        self.assertEqual(some_module.kwargs_only_func(1, bar=2), 123)
+        flexmock(some_module).should_receive("kwargs_only_func1").with_args(1, bar=2).and_return(
+            123
+        )
+        self.assertEqual(some_module.kwargs_only_func1(1, bar=2), 123)
 
     def test_mock_kwargs_only_func_fails_if_required_not_provided(self):
-        self.assertRaises(
-            MethodSignatureError,
-            flexmock(some_module).should_receive("kwargs_only_func").with_args,
-            1,
-        )
+        with assert_raises(
+            MethodSignatureError, 'kwargs_only_func1 requires keyword-only argument "bar"'
+        ):
+            flexmock(some_module).should_receive("kwargs_only_func1").with_args(1)
+        with assert_raises(
+            MethodSignatureError, 'kwargs_only_func2 requires keyword-only arguments "bar", "baz"'
+        ):
+            flexmock(some_module).should_receive("kwargs_only_func2").with_args(2)
 
 
 class TestReturnValue(unittest.TestCase):